chore(scripts): split for setup instructions from test cases instructions
Setup is moved to fork_setup.py
